Changing media id from URL determination
Fixes a Problem where the ID of the Twitter an opengraph images could not be determined from the URL, if the original image was larger than the desired size. This results in Yoast SEO returning the URL of a scaled version of the image, which could not be reverse resolved by the standard `attachment_url_to_postid` function. This was replaced by Yoast SEO's own `WPSEO_Image_Utils::get_attachment_by_url` which overcomes the problem.
Additionally the function `wpcom_vip_attachment_url_to_postid` had to be renamed, as the YoastSEO function uses this function resulting in a segmentation fault.
Fixes #60.